+++ crates/colab-lang-java/src/imports.rs
@@ -20,7 +20,7 @@ fn import_name<'a>(node: Node<'a>, source: &'a str) -> Option<(Node<'a>, String)
     // `asterisk` after `import com.x.*;`. We want a dotted name; for
     // `*` we return None so it never matches a literal target.
     for i in 0..node.named_child_count() {
-        let child = node.named_child(i)?;
+        let child = node.named_child(i as u32)?;
         match child.kind() {
             "scoped_identifier" | "identifier" => {
                 if let Ok(text) = child.utf8_text(source.as_bytes()) {
@@ -206,7 +206,7 @@ pub fn ensure(target: &str, source_code: &str) -> String {
     let root = tree.root_node();
     let mut insert_at: Option<usize> = None;
     for i in 0..root.named_child_count() {
-        let Some(child) = root.named_child(i) else {
+        let Some(child) = root.named_child(i as u32) else {
             break;
         };
         if child.kind() == "package_declaration" {
